Finish the whole project
Explanation:
There are 3 functions in the script element of the HTML :
1. getComputerChoice(), used to generate a random choice between Rock,
Paper, and Scissors. So the variable "randomNumber" would generate a
random number between 0-2. I'm using a switch as the conditional,
if it generated 0, "ROCK" string would be assigned to a variable called
"choice" that was initialized earlier outside the switch.
And if 1, "PAPER", and 2 for "SCISSORS". Variable "choice" would be
returned.

2. singleRun(playerSelection,computerSelection), used to compare the
results of the player selection and the computer selection. This is
where the game happens, this function works like a real life Rock, Paper
, Scissors.

3. game(), used to loop the game 5 times. This is where user inputs
their choices. The choices should be either rock, paper, or scissors,
other than that, the game would end tie. Capital letters would be
ignored as in whichever form the user input the choice, it would get
uppercased, so if user inputted "rOcK", it would get upercased to "ROCK"
.